Add a readOnly mode to HttpFields.Mutable.
This is useful for when a response is committed.  Taking an immutable from the response is good enough to write to the protocol, but the issue is that the mutable is left behind and can be further mutated, making request logging difficult.   This approach allows the mutations to be ignored once the readonly mode is engaged... and has the benefit of avoiding and array copy as well... at the cost of a boolean check on every mutation.