Moved client and server tests to new module jetty-http2-tests.
This was necessary also to overcome JPMS issues, in particular:
Tests in the http2-client module would require http2-common, but also http2-server, which also depends on http2-common.
Therefore, http2-client and http2-common would be in the module-path, but http2-server would be in the class-path.
This causes IllegalAccessExceptions because http2-server classes would not be able to access internal classes of http2-common (because http2-common does not export internal classes to the unnamed module, since http2-server is in the class-path).

Best solution was to move all tests to a new module, therefore all classes are in the class-path without JPMS problems.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>