Fix test-distribution classpath re resolver (#8187)
* Fix test-distribution classpath re resolver

As currently it seems mixed versions of resolver artifacts are
present on classpath. Resolver 1.8.0 introduces a binary
breakage.

Anyway, this fixes the thing, as impl was actually coming in
as transitive dep of maven-resolver-provider, that is NOT
the same version as the rest of resolver dependencies.

Relying on transitive while fixing other versions is wrong.

* Update to resolver 1.8.1

And add exclusion to banned dependency, that is used when
running in SISU container.