Fixes #8678 - Jetty client receives GO_AWAY and continue to send traffic on same connection
* Now upon receiving the GOAWAY, the connection is removed from the pool, so it cannot be used by new requests.
* HTTP2Session.removeStream() now happens _after_ notifying HEADERS and DATA events, although the Stream state change still happens before.
This is necessary to avoid that a "close" event is notified before a "headers" or "data" event.

With these changes, the race window of a client acquiring a connection while the server is closing it is reduced, but it is impossible to close it completely.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>
(cherry picked from commit fbc23ac6dc9ce2e7bcc5a17ea5b9ef38d8bd7b77)