Add Maven profile to automatically reformat code
Jetty code can now automatically be reformatted by running the following
command:

    mvn process-sources -Preformat

This will invoke the "reformat" Maven profile, which currently executes
the "impsort-maven-plugin" to organize imports, and the
"formatter-maven-plugin" to reformat Java source code, according to the
rules defined in build/build-resources/jetty-codestyle-eclipse-ide.xml.

Note that not all code is currently formatted according to these rules,
so a separate commit should incorporate the reformatting changes in the
future.

https://github.com/eclipse/jetty.project/issues/9302

Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>