Experiment with Conditional Handler Wrappers
ConditionalHandler0 is a simple combined approach with a single handler that has multiple different criteria that can be applied (method, URI, Inet) to determine if the wrapped handler should be applied or not.    There is also a skip parameter, which if true means the next handler is just skipped rather than the entire request not handled.  (Note that currently the Inet handling also has URI handling in its patterns - a real impl would just do one or the other URI handling).

Issue with ConditionalHandler0 is that it has a fixed set of conditions baked into it... so

ConditionalHandler1 is an attempt to make a composable ConditionalHandler.  It keeps the normal handler chain, but works out how far ahead it needs to jump over a chain of composed conditional handlers (it is currently always in skip mode).

ConditionalHandler2 is an attempt to make avoid composition by making a conditional branch.   The issue here is that any attributes of the branched handler (eg Graceful) need to be echoed onto the main handler.

ConditionalHandler3 is an attempt to make the branch explicit, so you don't need to echo Graceful.