Fixes #10482 - RewriteHandler with multiple HeaderPatternRules.
Reviewed the implementation of RewriteHandler, that was broken for those rules that were overriding `Rule.Handler.handle()`.

The problem was that the handling was not forwarded along the chain of rules, so only the last one was applied.

Removed (a breaking change) `Rule.Handler.handle(Response, Callback)` because it is necessary that the outermost request wrapper is passed to the `RewriteHandler` child `Handler`.
Updated implementations to override `handle(Request, Response, Callback)` instead, so that the request parameter is preserved along the chain.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>