Fixes #10482 - RewriteHandler with multiple HeaderPatternRules.
Updated the implementation after review.

Restored the `Rule.Handler.handle(Response, Callback)` method so that existing code won't break.

Now the wrapping at the constructor produces RH3(RH2(RH1(Request))), but the handling is performed from the innermost towards the outermost.
In this way, the order of rules is respected, both in the wrapping at rule application, and in the `Rule.Handler` handling.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>