Introduced IteratingCallback.abort(Throwable) to serialize calls to onCompleteFailure().
HttpSender now uses IC.abort() to serialize the abort of the request.

Fixed HTTP/2 and HTTP/3 to dispatch the sending of the request to another thread, to free the reader thread that read the server preface.
HTTP/1.1 does not need this, because just created connections do not need to read.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>