Removed non-Antora documentation.
Now the documentation has been migrated to Antora, so there is no need for the Java Asciidoctor plugin, and the directories that previously held the documentation using the Asciidoctor Maven plugin.

Updated the `javadoc-url` attribute with the correct URI.

Renamed the `version` attribute to `jetty-version`, and now using the POM value, rather than a hard-coded one.

Removed references to Asciidoctor from the main POM.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>