Fixed case when the trailer HEADERS frame is QPACK-blocked.
HTTP upper layers (core, servlet) may insist reading even if null was already returned to them because the trailer is QPACK-blocked.

Trying to read more was delegated to the QUIC layer, but for QUIC the stream is complete, since it has delivered the trailer HEADERS frame; the actual headers arrive on the HTTP/3 encoder stream, but it's a different stream.

Now trying to read more will return null until the stream is QPACK-unblocked.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>