Fixes #13375 - WebSocketClient attempts to use HTTP/2 on servers without SETTINGS_ENABLE_CONNECT_PROTOCOL
Now the client does not try to use HTTP/2 CONNECT + :protocol=websocket if the server does not support it, so the failure will happen in the client, not on the server.

When using `HttpClientTransportDynamic`, the client will try to fallback from HTTP/2 to HTTP/1 if possible, and retry the upgrade.

Moved the WebSocketOverHTTP2Test class to core, to avoid the eeN duplicates.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>