Fix `<license><name>` to SPDX in pom.xml
Automated license compliance checkers complain when the `<name>`
and the `<url>` tag of a `<license>` element in the pom.xml don't match.

Any mismatch requires manual investigation.

The current tag `<name>Apache Software License - Version 2.0</name>`
doesn't match the official license name because the official license
name is `Apache License` and not `Apache Software License`:
https://www.apache.org/licenses/LICENSE-2.0

There are two official ways to fill the `<name>` tag:
* The full legal name of the license as specified by https://maven.apache.org/ref/3.9.11/maven-model/maven.html
* The SPDX identifier as recommended by https://maven.apache.org/pom.html#Licenses

During code review the full legal name was rejected as wrong and it was requested that
the SPDX identifier must be used.

Therefore the `<name>` is changed from the full legal name to the SPDX identifier.
This fixes the issue with the wrong license name.

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>