Fixes StackOverflowError in redirect tests. (#13533)
A stack overflow was happening when a redirect was failed (for example, bad scheme of the redirect location).

This was caused as a side effect of the changes done in #13381.

HttpRedirector was registering a `Response.CompleteListener` that was notifying the original request/response listeners.
This is necessary to notify failures before the redirect request is sent.
But calling this listener after the redirect request is sent causes a stack overflow.

The solution is to check whether the redirect request is already part of the conversation.
If so, the notification of the listeners is performed by the `HttpClient` implementation.
Otherwise, it must be explicitly done by HttpRedirector.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>