Fixes #13583 - NPE in WriteFlusher after org.eclipse.jetty.http.BadMessageException: 417.
Fixes handling of an `Expect` header with an invalid value (one that is not 100-Continue).

For HTTP/1 replaced direct call to `badMessage()` with throwing, so that the request is not handled in `HttpConnection.onFillable()`.

HTTP/2 and HTTP/3 did not handle the case, so the handling was added.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>
(cherry picked from commit 391a544b0586c08779c470c222621be480c37447)
Signed-off-by: Simone Bordet <simone.bordet@gmail.com>