PR #13557 - fix bug in JettyWebSocketFrameHandler
The OnWebSocketClose notification can be called in two cases, if a close frame is received, or if the connection is notified as closed through FrameHandler.onClosed.

The callback should only be succeeded not failed if this is called twice, because this is a normal case.
Received a response close frame and OnWebSocketClose is called through JettyWebSocketFrameHandler.onFrame.
Then the connection is closed so JettyWebSocketFrameHandler.onClosed is called, but we cannot call OnWebSocketClose again.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>