Experiment with a map for associating objects with ByteBuffers
This map could be used to associate an object with a ByteBuffer for tunneling through APIs that only take ByteBuffer.  For example, and implementation supporting transferTo could associate a particular Content.Source with an empty buffer and write that to a Sink.  If the ultimate implementation of the sink takes the source and uses it, then no more writes are necessary, otherwise the content has to be written normally.