Fixes #13708 - Improve StateTrackingHandler reporting of incomplete callbacks.
When the `Handler` chain returns `false`, the request is not handled, and `Server` will respond with a 404.
In this case, we must cancel the task scheduled to check whether the `Handler` callback was completed,
since it will never be completed.
This avoids to report false positives for the "HandlerCallbackNotCompleted" event.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>